agentic-ai/src/agents/validator.py
"""
Placeholder for ResultValidator
"""
import logging
from typing import Dict, Any, List, Optional
from langchain_core.prompts import ChatPromptTemplate
from langchain_core.output_parsers import JsonOutputParser
from langchain_community.chat_models import ChatOpenAI # Using OpenAI as an example, needs to be configurable
import json

logger = logging.getLogger(__name__)

class ResultValidator:
    """
    A component that validates the results of executed subtasks
    against predefined criteria, potentially using an LLM.
    """
    def __init__(self, llm_model: Optional[Any] = None, validation_prompt: Optional[ChatPromptTemplate] = None, output_parser: Optional[JsonOutputParser] = None):
        self.llm = llm_model if llm_model else ChatOpenAI(temperature=0)
        self.validation_prompt = validation_prompt if validation_prompt else ChatPromptTemplate.from_messages(
            [
                ("system", "You are an expert result validator. Your task is to assess the outcome of a subtask against specific success criteria. Provide a 'status' (passed/failed), a 'score' (0.0-1.0), and 'feedback'. Output your answer as a JSON object."),
                ("human", "Subtask Result: {result}\nSuccess Criteria: {criteria}\nContext: {context}")
            ]
        )
        self.output_parser = output_parser if output_parser else JsonOutputParser()
        self.llm_chain = self.validation_prompt | self.llm | self.output_parser
        logger.info("Initialized ResultValidator with LLM-based validation.")

    def validate(self, result: Dict[str, Any], task_spec: Dict[str, Any], context: Dict[str, Any]) -> Dict[str, Any]:
        """
        Validates a single subtask result against the task's success criteria using an LLM.
        """
        logger.info(f"ResultValidator: Validating result for subtask: {result.get('subtask', {}).get('id', 'N/A')}.")
        
        success_criteria = task_spec.get("success_criteria", "The subtask should complete successfully.")
        
        try:
            llm_output = self.llm_chain.invoke({
                "result": json.dumps(result), # Pass result as JSON string to LLM
                "criteria": success_criteria,
                "context": context
            })

            # Ensure llm_output is a dictionary (from JsonOutputParser)
            if not isinstance(llm_output, dict):
                raise ValueError("LLM did not return a valid JSON object for validation.")

            status = llm_output.get("status", "failed").lower()
            score = float(llm_output.get("score", 0.0))
            feedback = llm_output.get("feedback", "No specific feedback provided by LLM.")

            return {
                "status": status,
                "score": score,
                "feedback": feedback,
                "validated_result": result,
                "criteria_met": status == "passed"
            }

        except Exception as e:
            logger.error(f"Error during LLM-based validation: {e}", exc_info=True)
            return {
                "status": "failed",
                "score": 0.0,
                "feedback": f"Validation failed due to internal error: {str(e)}",
                "validated_result": result,
                "criteria_met": False
            }