agentic-ai/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-ai-platform"
version = "0.1.0"
description = "Agentic AI Orchestration Platform with LangGraph and MCP"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langchain>=0.1.0",
    "langgraph>=0.0.30",
    "langchain-community>=0.0.10",
    "mcp>=0.1.0",
    "transformers>=4.35.0",
    "torch>=2.0.0",
    "trl>=0.7.0",
    "peft>=0.6.0",
    "datasets>=2.14.0",
    "mlflow>=2.8.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "docker>=6.1.0",
    "pydantic>=2.5.0",
    "networkx>=3.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0",
    "mypy>=1.5.0",
    "pylint>=3.0.0",
    "wandb>=0.15.0",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "ipykernel>=6.0",
    "pre-commit>=3.5.0",
]
gpu = [
    "torch>=2.0.0",
    "nvidia-cublas-cu12",
    "nvidia-cuda-nvrtc-cu12",
]
aws = [
    "boto3>=1.28.0",
    "sagemaker>=2.190.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/agentic-ai-platform"
Documentation = "https://github.com/yourusername/agentic-ai-platform#readme"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=xml"